requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.2.7
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
h11==0.14.0
html-parser==0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
Pillow==10.1.0
ply==3.11
protobuf==4.21.12
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.27.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
webdriver-manager==4.0.0
Werkzeug==3.0.1
wsproto==1.2.0
